requirements.txt
# Core
numpy==1.26.0
pandas==2.2.1
matplotlib==3.8.3
scikit-learn==1.5.0

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.13
deepface==0.0.89

# Web & GUI
flask==3.0.3
flask-socketio==5.3.6

# API & Utilities
google-api-python-client==2.123.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2

# Machine Learning
tf-keras==2.19.0
tensorflow==2.19.0
scipy==1.13.0